introduce PersistentModule mixin
