+++ rosys/config.py
@@ -1,23 +1,17 @@
-from dataclasses import dataclass
 from typing import Any
 
 from . import persistence
 
 
-@dataclass(slots=True, kw_only=True)
-class Config:
-    ui_update_interval: float = 0.1
-    simulation_speed: float = 1.0
-    needs_backup: bool = False
+class Config(persistence.PersistentModule):
 
-    def __post_init__(self) -> None:
-        persistence.register(self)
+    def __init__(self) -> None:
+        super().__init__()
+        self.ui_update_interval: float = 0.1
+        self.simulation_speed: float = 1.0
 
     def backup(self) -> dict[str, Any]:
         return {'simulation_speed': self.simulation_speed}
 
     def restore(self, data: dict[str, Any]) -> None:
         self.simulation_speed = data.get('simulation_speed', 1.0)
-
-    def invalidate(self) -> None:
-        self.needs_backup = True
