+++ rosys/persistence/__init__.py
@@ -1,0 +1,20 @@
+from dataclasses_json import Exclude, config
+
+from .converters import from_dict, replace_dataclass, replace_dict, replace_list, replace_set, to_dict
+from .persistent_module import PersistentModule
+from .registry import backup, restore
+
+exclude: dict[str, dict] = config(exclude=Exclude.ALWAYS)
+
+__all__ = [
+    'from_dict',
+    'replace_dataclass',
+    'replace_dict',
+    'replace_list',
+    'replace_set',
+    'to_dict',
+    'PersistentModule',
+    'exclude',
+    'backup',
+    'restore',
+]
